requirements.txt
alembic==1.0.3
amqp==2.4.0
aniso8601==4.0.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coverage==4.5.2
coveralls==1.5.1
DateTime==4.3
docopt==0.6.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.15.0
Flask-Migrate==2.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
Mako==1.0.7
MarkupSafe==1.1.0
passlib==1.7.1
psycopg2==2.7.6.1
PyJWT==1.7.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
redis==3.0.1
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
zope.interface==4.6.0
